@@ -7,20 +7,19 @@ homepage: https://github.com/OmkarDabade/markdown_editor_plus
 repository: https://github.com/OmkarDabade/markdown_editor_plus
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=3.1.5 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   expandable: ^5.0.1
-  markdown: ^6.0.0
-  flutter_markdown: ^0.6.12
-  font_awesome_flutter: ^10.2.1
+  markdown: ^7.2.2
+  flutter_markdown: ^0.7.1
+  font_awesome_flutter: ^10.7.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^4.0.0
 
 flutter: null
