@@ -1,7 +1,7 @@
 library markdown_editor_plus;
 
 export 'widgets/markdown_auto_preview.dart';
+export 'widgets/markdown_field.dart';
 export 'widgets/markdown_parse.dart';
 export 'widgets/markdown_toolbar.dart';
-export 'widgets/markdown_field.dart';
 export 'widgets/splitted_markdown_form_field.dart';
